refactor(prisma): restructure schema to add Ebook, Training, and Testimonial models
This commit introduces new models (Ebook, Training, and Testimonial) to better organize the data structure. The Item model has been refactored to remove redundant fields and establish relationships with the new models. Additionally, the database URL in the .env file has been updated to point to a new server. These changes improve maintainability and clarity of the schema.